requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
joblib==1.5.3
numpy==2.4.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
whitenoise==6.11.0
